refactor: change cassandra_cql to namespace module
Splitting out retry utility and retry policy from public
CassandraCqlManager class.
