+++ beaker_extensions/cassandra_cql/__init__.py
@@ -1,0 +1,4 @@
+from beaker_extensions.cassandra_cql.manager import CassandraCqlManager
+
+
+__all__ = ["CassandraCqlManager"]
