+++ beaker_extensions/cassandra_cql/manager.py
@@ -1,17 +1,15 @@
-from __future__ import absolute_import
-
-import logging
 import random
 import re
 import socket
-from functools import partial, wraps
-from typing import TYPE_CHECKING, Any, Callable, Protocol, TypeVar, cast, overload
-
 import six
-from beaker.exceptions import InvalidCacheBackendError, MissingCacheParameter
-from datadog import statsd
+from beaker_extensions.nosql import NoSqlManager
+from beaker_extensions.cassandra_cql.utils import retry
+from beaker_extensions.cassandra_cql.retry_policy import AlwaysRetryNextHostPolicy
+from beaker.exceptions import MissingCacheParameter
+from beaker.exceptions import InvalidCacheBackendError
 
-from beaker_extensions.nosql import Container, NoSqlManager
+import logging
+
 
 try:
     import cassandra
@@ -19,98 +17,12 @@ try:
     from cassandra.cluster import Cluster
     from cassandra.policies import (
         DCAwareRoundRobinPolicy,
-        RetryPolicy,
         TokenAwarePolicy,
     )
 except ImportError:
-    raise InvalidCacheBackendError(
-        "cassandra_cql backend requires the 'cassandra-driver' library"
-    )
-
-if TYPE_CHECKING:
-    from typing import Optional, Type, Union
-
+    raise InvalidCacheBackendError("cassandra_cql backend requires the 'cassandra-driver' library")
 
 log = logging.getLogger(__name__)
-
-
-class SupportsRetry(Protocol):
-    _tries = 0  # type: int
-    _RETRYABLE_EXCEPTIONS = tuple()  # type: tuple[Type[Exception], ...]
-
-
-F = TypeVar("F", bound=Callable[..., Any])
-Decorator = Callable[[F], F]
-
-# This signature is for setting the tries override
-@overload
-def _retry(
-    func=None,  # type: None
-    tries=None,  # type: Optional[int]
-):  # type: (...) -> Decorator
-    pass
-
-
-# This signature is for wrapping a function
-@overload
-def _retry(
-    func,  # type: F
-):  # type: (...) -> F
-    pass
-
-
-def _retry(
-    func=None,  # type: Optional[F]
-    tries=None,  # type: Optional[int]
-):  # type: (...) -> Union[F, Decorator[F]]
-    """_retry will call the decorated function a number of times if a registered Exception is raised
-
-    Args:
-        tries (int, optional): The number of times to try the function. Defaults to the Class attribute `_tries`.
-    """
-    if func is None:
-        return cast(Decorator[F], partial(_retry, tries=tries))
-
-    @wraps(func)
-    def wrapper(
-        self,  # type: SupportsRetry
-        *args,
-        **kwargs
-    ):
-        retry_count = tries or self._tries
-        _tries = retry_count
-        while _tries:
-            try:
-                return func(self, *args, **kwargs)
-            except self._RETRYABLE_EXCEPTIONS:
-                _tries -= 1
-                if not _tries:
-                    statsd.increment(
-                        "dd.cassandra_cql.retry",
-                        tags=[
-                            "function:{}".format(func.__name__),
-                            "status:error",
-                            "retry_count:{}".format(_tries),
-                        ],
-                    )
-                    raise
-                t = retry_count - _tries
-                statsd.increment(
-                    "dd.cassandra_cql.retry",
-                    tags=[
-                        "function:{}".format(func.__name__),
-                        "status:retry",
-                        "retry_count:{}".format(_tries),
-                    ],
-                )
-                log.warning(
-                    "Caught retryable exception on try=%d (stack "
-                    "trace below). Retrying.",
-                    t,
-                    exc_info=True,
-                )
-
-    return cast(F, wrapper)
 
 
 class CassandraCqlManager(NoSqlManager):
@@ -134,29 +46,15 @@ class CassandraCqlManager(NoSqlManager):
 
     connection_pools = {}
 
-    def __init__(
-        self,
-        namespace,
-        url=None,
-        data_dir=None,
-        lock_dir=None,
-        keyspace=None,
-        column_family=None,
-        **params
-    ):
-        NoSqlManager.__init__(
-            self, namespace, url=url, data_dir=data_dir, lock_dir=lock_dir, **params
-        )
+    def __init__(self, namespace, url=None, data_dir=None, lock_dir=None, keyspace=None, column_family=None, **params):
+        NoSqlManager.__init__(self, namespace, url=url, data_dir=data_dir, lock_dir=lock_dir, **params)
         connection_key = "-".join(
-            ["%r" % url, "%r" % keyspace, "%r" % column_family]
-            + ["%s:%r" % (k, params[k]) for k in params]
+            ["%r" % url, "%r" % keyspace, "%r" % column_family] + ["%s:%r" % (k, params[k]) for k in params]
         )
         if connection_key in self.connection_pools:
             self.db_conn = self.connection_pools[connection_key]
         else:
-            self.db_conn = _CassandraBackedDict(
-                namespace, url, keyspace, column_family, **params
-            )
+            self.db_conn = _CassandraBackedDict(namespace, url, keyspace, column_family, **params)
             self.connection_pools[connection_key] = self.db_conn
 
     def open_connection(self, host, port, **params):
@@ -190,9 +88,7 @@ class _CassandraBackedDict(object):
         cassandra.RequestExecutionException,
     )
 
-    def __init__(
-        self, namespace, url=None, keyspace=None, column_family=None, **params
-    ):
+    def __init__(self, namespace, url=None, keyspace=None, column_family=None, **params):
         if not keyspace:
             raise MissingCacheParameter("keyspace is required")
         if re.search(r"\W", keyspace):
@@ -212,9 +108,7 @@ class _CassandraBackedDict(object):
 
         consistency_level_param = params.get("consistency_level")
         if isinstance(consistency_level_param, six.string_types):
-            consistency_level = getattr(
-                cassandra.ConsistencyLevel, consistency_level_param.upper(), None
-            )
+            consistency_level = getattr(cassandra.ConsistencyLevel, consistency_level_param.upper(), None)
             if consistency_level:
                 self.__session.default_consistency_level = consistency_level
 
@@ -238,9 +132,7 @@ class _CassandraBackedDict(object):
         cluster_params["contact_points"] = contact_points[:2]
 
         if "max_schema_agreement_wait" in params:
-            cluster_params["max_schema_agreement_wait"] = int(
-                params["max_schema_agreement_wait"]
-            )
+            cluster_params["max_schema_agreement_wait"] = int(params["max_schema_agreement_wait"])
 
         # Clients should use any details they have to route intelligently
         if "datacenter" in params:
@@ -258,14 +150,12 @@ class _CassandraBackedDict(object):
         # retry policy that does. We don't want to use _only_ this because this
         # is only for timeouts from the cassandra coordinator's perspective,
         # and wouldn't retry if there was a failure reaching cassandra at all.
-        cluster_params["default_retry_policy"] = _NextHostRetryPolicy()
+        cluster_params["default_retry_policy"] = AlwaysRetryNextHostPolicy()
 
         # If username and password is provided in the params
         # then include an auth provider when connecting to the cluster
         if "username" in params and "password" in params:
-            auth = PlainTextAuthProvider(
-                username=params["username"], password=params["password"]
-            )
+            auth = PlainTextAuthProvider(username=params["username"], password=params["password"])
             cluster_params["auth_provider"] = auth
 
         log.info("Connecting to cassandra cluster with params %s", cluster_params)
@@ -336,7 +226,7 @@ class _CassandraBackedDict(object):
         )
         self.__del_stmt = self.__session.prepare(del_query)
 
-    @_retry
+    @retry
     def has_key(self, key):
         # NoSqlManager uses has_key() rather than `in`.
         rows = self.__session.execute(self.__contains_stmt, {"key": key})
@@ -344,7 +234,7 @@ class _CassandraBackedDict(object):
         assert count == 0 or count == 1
         return count > 0
 
-    @_retry(tries=2)
+    @retry(tries=2)
     def __setitem__(self, key, value):
         data = value if isinstance(value, bytes) else value.encode()
         if self._expiretime:
@@ -353,11 +243,9 @@ class _CassandraBackedDict(object):
                 {"key": key, "data": data, "[ttl]": self._expiretime},
             )
         else:
-            self.__session.execute(
-                self.__set_no_expire_stmt, {"key": key, "data": data}
-            )
+            self.__session.execute(self.__set_no_expire_stmt, {"key": key, "data": data})
 
-    @_retry
+    @retry
     def get(self, key):
         # NoSqlManager uses get() rather than [].
         #
@@ -375,7 +263,7 @@ class _CassandraBackedDict(object):
             pass
         return res.data
 
-    @_retry
+    @retry
     def __delitem__(self, key):
         self.__session.execute(self.__del_stmt, [key])
 
@@ -389,45 +277,3 @@ class _CassandraBackedDict(object):
             "keys, not the keys used by the rest of this class' api. "
             "Regardless, it doesn't seem to be used so I'm not implementing it."
         )
-
-
-class _NextHostRetryPolicy(RetryPolicy):
-    def on_read_timeout(
-        self,
-        query,
-        consistency,
-        required_responses,
-        received_responses,
-        data_retrieved,
-        retry_num,
-    ):
-        if retry_num == 0:
-            return self.RETRY_NEXT_HOST, None
-        else:
-            return self.RETHROW, None
-
-    def on_write_timeout(
-        self,
-        query,
-        consistency,
-        write_type,
-        required_responses,
-        received_responses,
-        retry_num,
-    ):
-        if retry_num == 0:
-            return self.RETRY_NEXT_HOST, None
-        else:
-            return self.RETHROW, None
-
-    def on_unavailable(
-        self, query, consistency, required_replicas, alive_replicas, retry_num
-    ):
-        if retry_num == 0:
-            return self.RETRY_NEXT_HOST, None
-        else:
-            return self.RETHROW, None
-
-
-class CassandraCqlContainer(Container):
-    namespace_class = CassandraCqlManager
